fastapi/requirements.txt
# Web framework
fastapi
uvicorn[standard]
gunicorn 

# Database
sqlalchemy
asyncpg
psycopg2-binary

# Config
python-dotenv
pydantic[email]

# HTTP Client
httpx

# Data
pandas
scikit-learn
joblib
dill
# MLflow for model loading
mlflow
boto3
# Optional: LLM integrations
langchain
langchain-community
langgraph
langsmith
openai
anthropic
sentence-transformers